move data scripts into data folder
