+++ .gitignore
@@ -1,9 +1,3 @@
-# folders
-/data
-/runs
-/results
-/models
-
 # Byte-compiled / optimized / DLL files
 __pycache__/
 *.py[cod]
@@ -113,9 +107,10 @@ venv.bak/
 #vscode
 .vscode/
 
-runs/
-results/
-models/
+# folders
+/runs
+/results
+/models
 
 .ipynb_checkpoints/
